Create apitran.py
This is the initial version of my APT toolkit, which I have dubbed apitran (Automatic PIcture TRANmission). In this current state, the script takes a .wav recording of a NOAA APT signal as input, and outputs a decoded image. The input file can be mono or stereo, at any samplerate.